requirements.txt
Pillow
tqdm
cairosvg
